migrate risk calculations to separate files
